+++ cloudevents/tests/test_http_events.py
@@ -25,7 +25,10 @@ from cloudevents.http import (
     CloudEvent,
     from_http,
     is_binary,
+    is_structured,
+    to_binary,
     to_binary_http,
+    to_structured,
     to_structured_http,
 )
 from cloudevents.sdk import converters
@@ -69,17 +72,13 @@ test_data = {"payload-content": "Hello World!"}
 app = Sanic(__name__)
 
 
-def post(url, headers, data):
-    return app.test_client.post(url, headers=headers, data=data)
-
-
 @app.route("/event", ["POST"])
 async def echo(request):
     decoder = None
     if "binary-payload" in request.headers:
         decoder = lambda x: x
     event = from_http(
-        request.body, headers=dict(request.headers), data_unmarshaller=decoder
+        dict(request.headers), request.body, data_unmarshaller=decoder
     )
     data = (
         event.data
@@ -91,25 +90,24 @@ async def echo(request):
 
 @pytest.mark.parametrize("body", invalid_cloudevent_request_body)
 def test_missing_required_fields_structured(body):
-    with pytest.raises(cloud_exceptions.CloudEventMissingRequiredFields):
+    with pytest.raises(cloud_exceptions.MissingRequiredFields):
         # CloudEvent constructor throws TypeError if missing required field
         # and NotImplementedError because structured calls aren't
         # implemented. In this instance one of the required keys should have
         # prefix e-id instead of ce-id therefore it should throw
         _ = from_http(
-            json.dumps(body),
-            headers={"Content-Type": "application/cloudevents+json"},
+            {"Content-Type": "application/cloudevents+json"}, json.dumps(body),
         )
 
 
 @pytest.mark.parametrize("headers", invalid_test_headers)
 def test_missing_required_fields_binary(headers):
-    with pytest.raises(cloud_exceptions.CloudEventMissingRequiredFields):
+    with pytest.raises(cloud_exceptions.MissingRequiredFields):
         # CloudEvent constructor throws TypeError if missing required field
         # and NotImplementedError because structured calls aren't
         # implemented. In this instance one of the required keys should have
         # prefix e-id instead of ce-id therefore it should throw
-        _ = from_http(json.dumps(test_data), headers=headers)
+        _ = from_http(headers, json.dumps(test_data))
 
 
 @pytest.mark.parametrize("specversion", ["1.0", "0.3"])
@@ -177,9 +175,9 @@ def test_roundtrip_non_json_event(converter, specversion):
     event = CloudEvent(attrs, compressed_data)
 
     if converter == converters.TypeStructured:
-        headers, data = to_structured_http(event, data_marshaller=lambda x: x)
+        headers, data = to_structured(event, data_marshaller=lambda x: x)
     elif converter == converters.TypeBinary:
-        headers, data = to_binary_http(event, data_marshaller=lambda x: x)
+        headers, data = to_binary(event, data_marshaller=lambda x: x)
 
     headers["binary-payload"] = "true"  # Decoding hint for server
     _, r = app.test_client.post("/event", headers=headers, data=data)
@@ -204,12 +202,12 @@ def test_missing_ce_prefix_binary_event(specversion):
         # breaking prefix e.g. e-id instead of ce-id
         prefixed_headers[key[1:]] = headers[key]
 
-        with pytest.raises(cloud_exceptions.CloudEventMissingRequiredFields):
+        with pytest.raises(cloud_exceptions.MissingRequiredFields):
             # CloudEvent constructor throws TypeError if missing required field
             # and NotImplementedError because structured calls aren't
             # implemented. In this instance one of the required keys should have
             # prefix e-id instead of ce-id therefore it should throw
-            _ = from_http(json.dumps(test_data), headers=prefixed_headers)
+            _ = from_http(prefixed_headers, json.dumps(test_data))
 
 
 @pytest.mark.parametrize("specversion", ["1.0", "0.3"])
@@ -226,7 +224,7 @@ def test_valid_binary_events(specversion):
             "ce-specversion": specversion,
         }
         data = {"payload": f"payload-{i}"}
-        events_queue.append(from_http(json.dumps(data), headers=headers))
+        events_queue.append(from_http(headers, json.dumps(data)))
 
     for i, event in enumerate(events_queue):
         data = event.data
@@ -247,7 +245,7 @@ def test_structured_to_request(specversion):
     data = {"message": "Hello World!"}
 
     event = CloudEvent(attributes, data)
-    headers, body_bytes = to_structured_http(event)
+    headers, body_bytes = to_structured(event)
     assert isinstance(body_bytes, bytes)
     body = json.loads(body_bytes)
 
@@ -267,7 +265,7 @@ def test_binary_to_request(specversion):
     }
     data = {"message": "Hello World!"}
     event = CloudEvent(attributes, data)
-    headers, body_bytes = to_binary_http(event)
+    headers, body_bytes = to_binary(event)
     body = json.loads(body_bytes)
 
     for key in data:
@@ -289,7 +287,7 @@ def test_empty_data_structured_event(specversion):
     }
 
     _ = from_http(
-        json.dumps(attributes), {"content-type": "application/cloudevents+json"}
+        {"content-type": "application/cloudevents+json"}, json.dumps(attributes)
     )
 
 
@@ -304,7 +302,7 @@ def test_empty_data_binary_event(specversion):
         "ce-time": "2018-10-23T12:28:22.4579346Z",
         "ce-source": "<source-url>",
     }
-    _ = from_http("", headers)
+    _ = from_http(headers, "")
 
 
 @pytest.mark.parametrize("specversion", ["1.0", "0.3"])
@@ -322,8 +320,8 @@ def test_valid_structured_events(specversion):
         }
         events_queue.append(
             from_http(
-                json.dumps(event),
                 {"content-type": "application/cloudevents+json"},
+                json.dumps(event),
             )
         )
 
@@ -344,7 +342,7 @@ def test_structured_no_content_type(specversion):
         "specversion": specversion,
         "data": test_data,
     }
-    event = from_http(json.dumps(data), {},)
+    event = from_http({}, json.dumps(data))
 
     assert event["id"] == "id"
     assert event["source"] == "source.com.test"
@@ -382,7 +380,7 @@ def test_cloudevent_repr(specversion):
         "ce-time": "2018-10-23T12:28:22.4579346Z",
         "ce-source": "<source-url>",
     }
-    event = from_http("", headers)
+    event = from_http(headers, "")
     # Testing to make sure event is printable. I could runevent. __repr__() but
     # we had issues in the past where event.__repr__() could run but
     # print(event) would fail.
@@ -398,5 +396,79 @@ def test_none_data_cloudevent(specversion):
             "specversion": specversion,
         }
     )
-    to_binary_http(event)
-    to_structured_http(event)
+    to_binary(event)
+    to_structured(event)
+
+
+def test_wrong_specversion():
+    headers = {"Content-Type": "application/cloudevents+json"}
+    data = json.dumps(
+        {
+            "specversion": "0.2",
+            "type": "word.found.name",
+            "id": "96fb5f0b-001e-0108-6dfe-da6e2806f124",
+            "source": "<my-source>",
+        }
+    )
+    with pytest.raises(cloud_exceptions.InvalidRequiredFields) as e:
+        from_http(headers, data)
+    assert "Found invalid specversion 0.2" in str(e.value)
+
+
+def test_invalid_data_format_structured_from_http():
+    headers = {"Content-Type": "application/cloudevents+json"}
+    data = 20
+    with pytest.raises(cloud_exceptions.InvalidStructuredJSON) as e:
+        from_http(headers, data)
+    assert "Expected json of type (str, bytes, bytearray)" in str(e.value)
+
+
+def test_wrong_specversion_to_request():
+    event = CloudEvent({"source": "s", "type": "t"}, None)
+    with pytest.raises(cloud_exceptions.InvalidRequiredFields) as e:
+        event["specversion"] = "0.2"
+        to_binary(event)
+    assert "Unsupported specversion: 0.2" in str(e.value)
+
+
+def test_is_structured():
+    headers = {
+        "Content-Type": "application/cloudevents+json",
+    }
+    assert is_structured(headers)
+
+    headers = {
+        "ce-id": "my-id",
+        "ce-source": "<event-source>",
+        "ce-type": "cloudevent.event.type",
+        "ce-specversion": "1.0",
+        "Content-Type": "text/plain",
+    }
+    assert not is_structured(headers)
+
+
+def test_empty_json_structured():
+    headers = {"Content-Type": "application/cloudevents+json"}
+    data = ""
+    with pytest.raises(cloud_exceptions.InvalidStructuredJSON) as e:
+        from_http(
+            headers, data,
+        )
+    assert "Failed to read fields from structured event. " in str(e.value)
+
+
+def test_uppercase_headers_with_none_data_binary():
+    headers = {
+        "Ce-Id": "my-id",
+        "Ce-Source": "<event-source>",
+        "Ce-Type": "cloudevent.event.type",
+        "Ce-Specversion": "1.0",
+    }
+    event = from_http(headers, None)
+
+    for key in headers:
+        assert event[key.lower()[3:]] == headers[key]
+    assert event.data == None
+
+    _, new_data = to_binary(event)
+    assert new_data == None